Speech Emotion Recognition (SER) Using Deep Learning
This project is part of the final Data Mining project for the ITC Fellow Program 2020. The goal of this project is to build a machine learning pipeline for Speech Emotion Recognition (SER), a task that involves classifying emotions based on speech. The project utilizes deep learning techniques, specifically convolutional neural networks (CNNs) and transfer learning, to analyze audio signals and predict emotional states from the speech.